Preliminary version of validation sanity check.
To validate my parser, I wrote logic that takes in the AST representation and pretty-prints it back.
I then parse that representation, pretty-print the re-parsed AST, and verify that all of them match.
I need to clean up the code a bit, but the logic is all here.
The logic is all in `parse_dsl.py`.
`ast_test.py` only has some small utilities.
The changes I made in the DSL are to facilitate parsing, rather than semantic changes.