First pass at mutating games
`mutation_experiments.ipynb` visualizes some of the games that I skimmed through.
`mutation_experiments.py` gives a hacky/hardcoded approach to mutating games.
It mutates by replacing one of the functions under a `(then)` operator with one from another game, and then fixing variable names.
Fixing variable names is actually non-trivial, but I think the current solution works.
To visualize these, so we can reason about them, I added support to print to HTML in `ast_printer.py`, and also to mark the old/new mutated functions clearly.

There are two issues I see with generated games right now.
One is that this substitution often creates two `(once)` clauses in a row, which is rather unnatural -- it's something I rarely if ever wrote down translating myself.
The second is the objects we currently don't quantify over.
We disucssed a solution for that -- quantify over them in the raw representation and pre-process for a given room with instantiating with singleton items -- that I need to implement.