First pass on interactive version games + reusing preferences with different objects.
Added the ~20 games from participants in the interactive versions that are valid and not underconstrained.
This also required writing a data loder form the Firestore data (in the notebook).
And modifying some of the code utils, including the printer.
Finally, added an initial take on how to reuse preferences with different objects.
This merits further discussion.