Preliminary version of code-generated PDF documentation
Finished a version of the code that generates a doc from code.
It's a combination of analyzing the DSL examples we already have with hard-coded documentation for each syntax element.
I use the data from the DSL examples to decide what to render and how to render it.
However, much of the documentation is hard coded for now.