Third pass at _on
This implementation projects a point slightly below the bottom of the upper object and checks to see if the point is contained in the bounding box of the lower object.

The problem is in cases where the upper object is resting on the lower object in such a way that its center of mass is over air (e.g. a plank of wood resting on two pillars spaced apart)