Updates to evolutionary sampler
Fixed a bug inthe RegrowthSampler where under particular conditions
(a parseinfo recurring in a game due to crossing over and making multiple copies?)
it could fail to find the correct parent and then throw an error.

Changed the PopulationBasedSampler to keep an energy model and fitness featurizer separately (as the MCMC does) so it can easily visualize samples.

Added tracking of the population fitness stddev to PopulationBasedSampler

Added a crossover operator for full sections.

Cleaned up the use of command line args a bit.